Refactoring: Moved Environment flags to EnvironmentHax. (Environment flags are independent from the Level editor)
